% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{density_estimation}
\alias{density_estimation}
\title{LocPoly Density estimation on complicated domains}
\usage{
density_estimation(
  data,
  domain,
  at_points = NULL,
  resolution = NULL,
  bandwidth = NULL,
  degree = 3
)
}
\arguments{
\item{data}{The data as a \code{ppp} object}

\item{domain}{A window of class \code{owin}}

\item{at_points}{The points}

\item{resolution}{Not used}

\item{bandwidth}{The multivariate bandwidth as a named vector \code{c(x=hx, y=hy)}}

\item{degree}{The degree of the local polynomial approximation}
}
\value{
An estimation of the density of the \code{data} (on \code{domain}) computed at \code{at_points}
}
\description{
LocPoly Density estimation on complicated domains
}
