% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{points_inside_owin}
\alias{points_inside_owin}
\title{Select the points that belong to a window}
\usage{
points_inside_owin(points, W)
}
\arguments{
\item{points}{The points from which we select as a ppp object}

\item{W}{The window as an owin object}
}
\value{
The selected points
}
\description{
Select the points that belong to a window
}
