% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_sector.R
\name{polynomial_sector}
\alias{polynomial_sector}
\title{Polynomial sector Window}
\usage{
polynomial_sector(k, npoly = 128)
}
\arguments{
\item{k}{Degree of the polynomial sector}

\item{npoly}{number of points (resolution) in the polynomial line (boundary)}
}
\value{
An object of class "owin" specifying a window
}
\description{
Polynomial sector Window
}
\examples{
plot(polynomial_sector(1.8))

}
