% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regular_points_in_owin}
\alias{regular_points_in_owin}
\title{Construct a regular grid of n x n points in a window}
\usage{
regular_points_in_owin(W, n)
}
\arguments{
\item{W}{A window of type owin}

\item{n}{The number of points in the enclosing rectangle is n x n.}
}
\value{
The regular grid as an object of type ppp
}
\description{
Construct a regular grid of n x n points in a window
}
