% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_poly}
\alias{risk_poly}
\alias{risk_norm}
\title{Risk associated to polynomial densities on polynomial sector}
\format{
\subsection{\code{risk_poly}}{

A data frame with 960,000 rows and 6 columns:
\describe{
\item{rep}{Replication number}
\item{method}{Estimation method: sparr, lp0, lp1}
\item{k}{Degree of the polynomial sector}
\item{n}{Number of observed data}
\item{h}{Bandwidth}
\item{value}{Value of the squared error between f(0,0) and the estimation}
}
}

\subsection{\code{risk_norm}}{

A data frame with 960,000 rows and 6 columns.
Normal densities
}
}
\usage{
risk_poly

risk_norm
}
\description{
Simulated data
TBC
}
\keyword{datasets}
