% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_sector.R
\name{simulate_risk_polysec}
\alias{simulate_risk_polysec}
\title{Title}
\usage{
simulate_risk_polysec(NN, RR, HH, KK, density_type)
}
\arguments{
\item{NN}{iterator - nb of observations}

\item{RR}{iterator -  replications}

\item{HH}{iterator - family of bandwidth}

\item{KK}{iterator - degree poly sector}

\item{density_type}{norm or poly}
}
\value{
a tibble of simulated risks for sparr and LP
}
\description{
Title
}
